requirements.txt
Django~=5.0.6
aiogram~=3.6.0
python-dotenv~=1.0.1
asgiref~=3.8.1
requests~=2.31.0